requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.20.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
lxml==5.1.0
multidict==6.0.5
psycopg==3.1.18
psycopg-binary==3.1.8
PyYAML==6.0.1
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==2.2.1
yarl==1.9.4
